+++ OneBlink.SDK/models/FormsAppEnvironmentStylesButtons.cs
@@ -1,0 +1,31 @@
+namespace OneBlink.SDK.Model
+{
+    public class FormsAppEnvironmentStylesButtons
+    {
+        public FormsListStylesButton submit
+        {
+            get; set;
+        }
+        public FormsListStylesButton cancel
+        {
+            get; set;
+        }
+        public FormsListStylesButton saveDraft
+        {
+            get; set;
+        }
+        public FormsListStylesButton cancelPromptYes
+        {
+            get; set;
+        }
+        public FormsListStylesButton cancelPromptNo
+        {
+            get; set;
+        }
+    }
+
+    public class FormsListStylesButtons : FormsAppEnvironmentStylesButtons
+    {
+    }
+}
+